requirements.txt
chromium-compact-language-detector
twython>=3.1.2
pyyaml
